combined api and auth user models
