+++ auth_backend/modules/user/serializers.py
@@ -1,8 +1,8 @@
-from .models import User
+from .models import BaseVologUser
 from rest_framework import serializers
 
 
 class UserSerializer(serializers.ModelSerializer):
     class Meta:
-        model = User
+        model = BaseVologUser
         fields = '__all__'
