+++ config/settings.py
@@ -104,7 +104,7 @@ ACCOUNT_USERNAME_REQUIRED = False
 ACCOUNT_AUTHENTICATION_METHOD = 'email'
 ACCOUNT_EMAIL_REQUIRED = True
 
-AUTH_USER_MODEL = 'user.User'
+AUTH_USER_MODEL = 'user.BaseVologUser'
 
 # Password validation
 # https://docs.djangoproject.com/en/3.1/ref/settings/#auth-password-validators
